requirements.txt
flask
requests
google-cloud-language
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn
